+++ trpc/routes/stores/list/route.ts
@@ -118,7 +118,7 @@ export const listActiveStoresProcedure = publicProcedure
         const reviewStats = await db
           .select({
             storeId: reviews.storeId,
-            averageRating: sql<number>`ROUND(AVG(${reviews.rating}), 1)`.as("averageRating"),
+            averageRating: sql<number>`ROUND(AVG(${reviews.rating})::numeric, 1)`.as("averageRating"),
             reviewCount: sql<number>`COUNT(*)`.as("reviewCount"),
           })
           .from(reviews)
